requirements.txt
# Core ML with CPU-only PyTorch
torch==2.8.0+cpu
transformers>=4.56.0,<4.58.0
sentence-transformers==5.1.1

# Keep all your other packages as-is
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
bitarray==3.7.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
fastapi==0.119.0
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pdfreader==0.1.15
pillow==11.3.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.0
pydantic_core==2.41.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
slack_bolt==1.26.0
slack_sdk==3.37.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
sendgrid==6.9.7